Mini-twotech - An in-game crafting guide
Replacing the in-game tab menu, showing Uses- and Produces- transitions of holding or pointer-hitting item. Transitions with nearby actor or target are prioritized to be displayed higher up on the list. Hover on the transition the actor or target nearby will be highlighted (since the hintarrow is not on THOL yet)