Adjust fertile age to 15, so Eves have 1 minute to say NO BB
Hesitated to make infertility a setting. It may promote solo play a bit too much and completely removes the possibility of those stories of solo player having a wonderful helpful accidental child

May still do it in the future, but with message to encourage players to try more parenting gameplay. Other tips/messages are planned