When contained object decays, execute both out- and in-transitions, in that order
Think of the decay transition of the contained object as taking out the old object (which may change the container) and then putting the new object in the potentially new container.

Also allow the container to go through 1-sec decay immediately after the out-transition. (A bit of a "hard code" fix for how table switches between large- and small-slot versions...)